Fixed ResourceAsStream Dependencies
Changed the internal file structure so .css, .js, and .html files would
be included in the jar file in a place so that -mopsreport will work as
an additional argument to the jar file and from within Eclipse.
Previously it would run in Eclipse but not as a standalone jar file.